+++ src/main/java/hotplace/repository/HotPlaceRepository.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 
 public interface HotPlaceRepository {
 
-    int save(Long memberId, Long contentId, HotPlace hotPlace);
+    int save(Long memberId, int contentId, HotPlace hotPlace);
 
     Optional<HotPlace> findById(Long hotPlaceId);
 }
